completed product controller and dtos for brand and category.

+++ src/main/java/com/example/order_manager_api/controllers/CategoryController.java
@@ -1,39 +1,46 @@
 package com.example.order_manager_api.controllers;
 
+import com.example.order_manager_api.dto_assemblers.CategoryDtoAssembler;
+import com.example.order_manager_api.dtos.CategoryDto;
 import com.example.order_manager_api.models.Category;
 import com.example.order_manager_api.services.ICategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+import java.util.Collection;
 
 @RestController
 public class CategoryController {
     @Autowired
     private ICategoryService categoryService;
+    @Autowired
+    private CategoryDtoAssembler categoryDtoAssembler;
 
     @GetMapping("/v1/categories")
-    public List<Category> findAll(){
-        return categoryService.findAll();
+    public Collection<CategoryDto> findAll(){
+        return categoryDtoAssembler.toCollection(categoryService.findAll());
     }
+
     @GetMapping("/v1/categories/{id}")
-    public Category findById(@PathVariable Integer id){
-        return categoryService.findById(id);
+    public CategoryDto findById(@PathVariable Integer id){
+        return categoryDtoAssembler.toDto(categoryService.findById(id));
     }
 
     @PostMapping("/v1/categories")
-    public Category save(@RequestBody Category newCategory){
-        return categoryService.save(newCategory);
+    public CategoryDto save(@RequestBody Category newCategory){
+        return categoryDtoAssembler.toDto(categoryService.save(newCategory));
     }
 
     @PutMapping("/v1/categories/{id}")
-    public Category update(@PathVariable Integer id,@RequestBody Category newCategory){
-        return categoryService.update(id,newCategory);
+    public CategoryDto update(@PathVariable Integer id,@RequestBody Category newCategory){
+        return categoryDtoAssembler.toDto(categoryService.update(id,newCategory));
     }
 
     @DeleteMapping("/v1/categories/{id}")
-    public void delete(@PathVariable Integer id){
-        if(findById(id) != null)
-            categoryService.delete(id);
+    public ResponseEntity<?> delete(@PathVariable Integer id){
+        if(findById(id) != null) categoryService.delete(id);
+
+        return ResponseEntity.noContent().build();
     }
 }
